.....J.... [ZBX-4663] add support for dots in attribute name
In JMX monitoring to get the value of a monitored item we specify
object name and attribute name. If the value is composite we specify
also the field name. The composite data can be nested so a real
attribute name and fields are separated by dots. E. g.

  jmx[com.example:Type=mytest,myattribute.compData.field1]

But sometimes an attribute name or a composite data field name can
contain dot. Before this fix it was not possible to monitor such
item.

This fix enables support for dots in attribute name, which just need
to be escaped with a backslash. E. g.

  jmx[com.example:Type=mytest,my\.attribute.comp\.data.field\.1]