...G...... [ZBX-6140] Fixed job creation for script execution on windows 2000            [merge ^/branches/2.0 -c r33379]
Windows 2000 does not support JobObjectExtendedLimitInformation command for
SetInformationJobObject() function. This was used for two things:
1) to clear JOB_OBJECT_LIMIT_BREAKAWAY_OK flag so the new processes can't
   disassociate themselves from parent job,
2) set JOB_OBJECT_LIMIT_KILL_ON_JOB_CLOSE flag so that all child processes
   are terminated when the job handle is closed.

In the first case the flag is not set by default, so we don't have to clear it.

In the second case the child processes are explicitly terminated when
TerminateJobObject() is called, so termination of children when job handle
is closed is irrelevant and therefore this flag can be omitted.