.......PS. [DNSTEST-1] only consider AUTHORITY section when querying name server
We decided to ignore ANSWER section and always query non-existent
domain. The logic is as follows:

1. Query the local resolver for the key set (DNSKEY)
  A. Validate AD bit is present
2. Query the TLD's name servers (from the template), one by one,
requesting an A record for PREFIX.TLD, and for each one validate:
  A. Status code should be NXDOMAIN or NOERROR
  B1. If EPP is enabled, validate:
    i. The response most be a referral (i.e. The AUTHORITY section
should contain at least one NS RR for the last label in PREFIX,
e.g. "icann-test.icanntest1" when querying for
"blahblah.icann-test.icanntest1")
    ii. Timestamp, extracting it from the NS RRs in the AUTHORITY
  section
    iii. If DNSSEC is enabled, validate all DS RRSIGs in the AUTHORITY
section against the appropriate key obtained in step 1
  B2. If EPP is disabled and DNSSEC is enabled
    i. Validate all NSEC and NSEC3 RRSIGs in the AUTHORITY section
against the appropriate key obtained in step 1
3. Calculate RTT, etc.