...G...... [ZBX-2966] treat negative available filesystem size as 0
Available filesystem size can be negative if a space reserved for
non-privileged users is hit. We treat it as zero as for a user it
makes filesystem monitoring much easier.

The negative value may be stored in an unsigned integer which makes it
harder to identify it. We use a special macro IS_NEGATIVE() which
checks the top bit of a value.